CVE-2023-31346 linux-firmware cbr-7.9
Please note that the dist tags for the source rpm and the base binary rpm
are different from all the others.  I suspect a mistake in the spec file,
but, since I came upon this well after it was built, making sure the CSAF
is accurate seems to be a better choice than doing another build.

Signed-off-by: Jonathan Dieter <jdieter@ciq.com>