CVE-2023-44487 pki-servlet-engine LTS 8.6, 8.8, 9.2 and 9.4
For 8.8, 9.2 and 9.4, pki-servlet-engine is obsolete and automatically
replaced by tomcat, so we're marking as known_not_affected.

For 8.6, there is no fix, and, due to the fact that this is a module, no
way to backport the tomcat fixes into `pki-servlet-engine`.  Therefore,
we are marking it as no_fix_planned.

Signed-off-by: Jonathan Dieter <jdieter@ciq.com>