CVE-2005-2541 tar CBR 7.9 - WONTFIX
This CVE is debatable.  Basically, the CVE is that tar doesn't give a
warning when extracting setuid or setgid files, but the user has to pass
the `-p` flag in in order for the files to actually be setuid or setgid.

Red Hat has closed this as WONTFIX, and all Debian releases are stil
vulnerable after almost 20 years, so we're in a strong position to set this
as WONTFIX.

Signed-off-by: Jonathan Dieter <jdieter@ciq.com>