CVE-2024-30202 emacs CBR 7.9 - Not affected
According to the CVE, "The Emacs package, as shipped in Red Hat Enterprise
Linux 8 and 9, is not affected by this vulnerability because the vulnerable
code was introduced in a newer version of Emacs."

Since RHEL 8 and 9 are too old to be affected, CentOS 7 is definitely too
old!  To verify, I’ve checked the code and the vulnerable file doesn’t
exist in CentOS 7.

Signed-off-by: Jonathan Dieter <jdieter@ciq.com>