[LTS 8.6] Add CVEs fixed prior to CSAF project
These CVEs were completed prior to the CSAF project getting into full
swing.  We know via release tagging in the kernel-src-tree which
binaries they're associated.  We picked CSAF files if they had the same
NVR (save some minor variation) to get the all the binary file names. If
we had a release that does not seem to be represented in CSAF we picked
the nearest one that existed and used those binary file names.

Kernels with NVR 4.18.0-372.32.1.0.5 we could match.
Kernels with NVR 4.18.0-372.32.1.0.6 we could not and used a newer kernel NVR:
 kernel-4.18.0-372.32.1.el8_6.86ciq_lts.8.1

The original binary list is no longer accessible to have a 1:1 match