[LTS 8.8] Add CVEs fixed prior to CSAF project
These CVEs were completed prior to the CSAF project getting into full
swing.  We know via release tagging in the kernel-src-tree which
binaries they're associated.  We picked CSAF files if they had the same
NVR (save some minor variation) to get the all the binary file names. If
we had a release that does not seem to be represented in CSAF we picked
the nearest one that existed and used those binary file names.